Adding reservoir time remaining estimate
This is obviously hard to predict, and it's not that useful. When `reservoir_days_remaining > 1`, the `reservoir_minutes_remaining` is noise, and isn't correlated with the day boundary like you'd expect. Once `reservoir_days_remaining == 1`, the `reservoir_minutes_remaining` correlates with the "Time left: HH:MM" display on the pump status screen.

When the status screen says "Time left: --:--", all three bytes are 0.

Here's my readings as the time left (suddenly) approaches 0:
```
01053a, 010534, 010514, 01050e, 010509, 010504, 0104ff, 0104fa, 0104f5, 0104f0, 0104eb, 0104e6, 0104e1, 0104dc, 0104cd, 0104c8, 010055, 010050, 000000
```